Audit Wave C: every stamp says the rule it applied
The math audit's stamp-honesty batch. The alpha-denominator convention
had one string and two averaging rules behind it: wholeCorpusMean-v1
now means exactly what the ResidualNormTally computes (every measured
position counts once), and the rule the sidecar writers actually apply
— each text contributes the mean over its own reading window, texts
averaged with equal weight — is minted as perTextMean-v1, on both
engines, with the ambiguous CURRENT constant removed so no future
writer can stamp without choosing. Existing wholeCorpusMean-v1
sidecars are grandfathered reader-side as the per-text rule (no tally
number has ever reached a sidecar on either engine), nothing is
rewritten, and the missing discriminating fixture now exists: variable
-length texts at a pooled reading where the two rules produce 6.0 and
4.0 through real code paths, plus the lastToken case proving they
coincide where the grandfathering assumes it. A propagation bug found
in passing: the Swift rewrap of a multi-concept extraction re-defaulted
the stamp, so a bank-denominated grand-mean vector would have been
filed under the per-text rule — the exact dishonesty F1 names — now
carried and pinned (F1).

Swift's mean paths accumulate in Double and cast once, removing the
sequential-float32 ceiling on cross-engine grand-mean agreement; every
committed parity fixture passed unchanged, so the motion stayed inside
every tolerance and no fixture needed regenerating (note 16). And the
neutral token-bank draw stops leaning on CPython's unguaranteed
random.sample stability: a line-for-line SplitMix64 Fisher-Yates twin
of the Swift downsampler, same seed derivation, same draw, pinned by
shared integer literals — pinned bases keep verifying because
neutralPCBasisHash travels by bytes; the default cap deliberately
stays engine-local (Swift's 4096 sits behind a memory preflight the
server does not have), and a cap of zero now selects nothing instead
of banking the whole corpus (note 9). PORTABILITY-CONTRACTS gains the
section the bank never had.

Verified personally: stamp definitions, grandfather rule, and the
rewrap fix read in full; the token-bank literals fired end-to-end
(seed and draw match the pinned values digit for digit). Both suites
green under my own invocation (Swift TEST SUCCEEDED; Python 5063
passed, 9 skipped), public scan clean.

+++ Sources/SteeringKit/Extraction/SteeringVectorMath.swift
@@ -375,18 +375,33 @@ public enum SteeringVectorMath {
     }
 
     /// Element-wise mean of equal-length rows.
+    ///
+    /// **Accumulated in `Double`, returned as `[Float]`** (2026-08-28 audit,
+    /// convention note 16). The rows are float32 — the deliberate cross-engine
+    /// storage dtype, unchanged — but a SEQUENTIAL float32 sum accrues error
+    /// on the order of N·ε, while numpy's `ndarray.mean` sums pairwise at
+    /// ~log₂(N)·ε. Over a few hundred stimulus rows that difference was the
+    /// ceiling on cross-engine agreement for a re-derived grand mean
+    /// (~1e-4..1e-3 relative in unlucky cases), which is a numerics artifact
+    /// masquerading as a recipe divergence. Accumulating in `Double` and
+    /// casting ONCE at the end costs nothing here (these run over small
+    /// captured activations on the CPU) and moves this engine's value TOWARD
+    /// the exact mean — and therefore toward numpy's. `ConceptExtractor
+    /// .neutralMeanPerLayer` has always accumulated this way; this brings the
+    /// mean / meanDifference / grandMeanDifference paths, and the PCA
+    /// centring that also calls through here, into line with it.
     public static func mean(_ rows: [[Float]]) throws -> [Float] {
         guard let first = rows.first else { throw SteeringVectorError.emptyInput }
-        var sum = [Float](repeating: 0, count: first.count)
+        var sum = [Double](repeating: 0, count: first.count)
         for row in rows {
             guard row.count == first.count else {
                 throw SteeringVectorError.dimensionMismatch(
                     expected: first.count, found: row.count)
             }
-            for i in row.indices { sum[i] += row[i] }
+            for i in row.indices { sum[i] += Double(row[i]) }
         }
-        let n = Float(rows.count)
-        return sum.map { $0 / n }
+        let n = Double(rows.count)
+        return sum.map { Float($0 / n) }
     }
 
     /// CAA direction: mean(positive activations) − mean(negative activations).
